Added Categories
made a categories with the last tab. moved the upper tab from home to categories.
