+++ src/Home.js
@@ -1,71 +1,6 @@
 import styles from './Style.js';
 import React from 'react';
 import { StyleSheet, Text, View, Image, Button } from 'react-native';
-import { TabNavigator } from "react-navigation";
-
-class Featured extends React.Component{
-  render(){
-    return(
-      <View>
-        <Text>Featured</Text>
-      </View>
-    );
-  }
-}
-
-class Women extends React.Component{
-  render(){
-    return(
-      <View>
-        <Text>Women</Text>
-      </View>
-    );
-  }
-}
-
-class Men extends React.Component{
-  render(){
-    return(
-      <View>
-        <Text>Men</Text>
-      </View>
-    );
-  }
-}
-
-class Children extends React.Component{
-  render(){
-    return(
-      <View>
-        <Text>Children</Text>
-      </View>
-    );
-  }
-}
-
-const SubTabNav = TabNavigator({
-    Featured: {
-      screen: Featured
-    },
-    Women: {
-      screen: Women
-    },
-    Men: {
-      screen: Men
-    },
-    Children: {
-      screen: Children
-    },
-},
-    {
-      tabBarPosition: 'top',
-      animationEnabled: false,
-      swipeEnabled: true,
-      tabBarOptions: {
-        activeTintColor: '#570761',
-      },
-
-});
 
 class HomeScreen extends React.Component {
   static navigationOptions = {
@@ -79,8 +14,7 @@ class HomeScreen extends React.Component {
   };
   render() {
     //const { navigate } = this.props.navigation;
-    return <SubTabNav />;
-      /*
+    return (
       <View style={styles.container}>
         <Text>Welcome to Shwop! OWIEURWLI </Text>
         <Image
@@ -88,8 +22,7 @@ class HomeScreen extends React.Component {
           style={{width: 333, height: 130}}
         />
       </View>
-      */
-    //);
+    );
   }
 }
